+++ e2e/runner/admin_evm.go
@@ -1,8 +1,6 @@
 package runner
 
 import (
-	"fmt"
-
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/stretchr/testify/require"
 
@@ -14,7 +12,7 @@ func (r *E2ERunner) UpdateTssAddressForConnector() {
 
 	tx, err := r.ConnectorEth.UpdateTssAddress(r.EVMAuth, r.TSSAddress)
 	require.NoError(r, err)
-	r.Logger.Info(fmt.Sprintf("TSS Address Update Tx: %s", tx.Hash().String()))
+	r.Logger.Info("TSS Address Update Tx: %s", tx.Hash().String())
 	receipt := utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, tx, r.Logger, r.ReceiptTimeout)
 	utils.RequireTxSuccessful(r, receipt)
 
@@ -28,7 +26,7 @@ func (r *E2ERunner) UpdateTssAddressForErc20custody() {
 
 	tx, err := r.ERC20CustodyV2.UpdateTSSAddress(r.EVMAuth, r.TSSAddress)
 	require.NoError(r, err)
-	r.Logger.Info(fmt.Sprintf("TSS ERC20 Address Update Tx: %s", tx.Hash().String()))
+	r.Logger.Info("TSS ERC20 Address Update Tx: %s", tx.Hash().String())
 	receipt := utils.MustWaitForTxReceipt(r.Ctx, r.EVMClient, tx, r.Logger, r.ReceiptTimeout)
 	utils.RequireTxSuccessful(r, receipt)